Split most functions into new JS files for easier maintenance.
