+++ node/risk-app/server/src/js/map-page/scenario-bars.js
@@ -16,6 +16,18 @@ function scenarioDisplayUpdate (scenarioBar) {
   })
 }
 
+function handleScroll (scenarioBar, arrows) {
+  scenarioBar.addEventListener('scroll', function () {
+    const currentScrollPosition = scenarioBar.scrollLeft
+    const divWidth = scenarioBar.offsetWidth
+    const scrollWidth = scenarioBar.scrollWidth
+
+    arrows[0].classList.toggle('hide', currentScrollPosition === scrollWidth - divWidth)
+    arrows[1].classList.toggle('hide', currentScrollPosition === 0)
+  })
+}
+
 module.exports = {
-  scenarioDisplayUpdate
+  scenarioDisplayUpdate,
+  handleScroll
 }
